Backend/requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
numpy==1.20.1
pip==21.0.1
pytz==2021.1
scipy==1.6.1
setuptools==54.1.0
sqlparse==0.4.1
